+++ Block/Info.php
@@ -200,19 +200,19 @@ class Info extends ConfigurableInfo
         $data = $info->getSfcoFirstPayment();
         if (!empty($data)) {
             $data = unserialize($data);
-            $date = preg_replace('/^([0-9]{2})([0-9]{2})([0-9]{4})$/', '$1/$2/$3', $data['date']);
+            $date = (isset($data['date']) ? preg_replace('/^([0-9]{2})([0-9]{2})([0-9]{4})$/', '$1/$2/$3', $data['date']) : 'N/A');
             $result['first'] = sprintf('%s (%s)', $data['amount'] / 100.0, $date);
         }
         $data = $info->getSfcoSecondPayment();
         if (!empty($data)) {
             $data = unserialize($data);
-            $date = preg_replace('/^([0-9]{2})([0-9]{2})([0-9]{4})$/', '$1/$2/$3', $data['date']);
+            $date = (isset($data['date']) ? preg_replace('/^([0-9]{2})([0-9]{2})([0-9]{4})$/', '$1/$2/$3', $data['date']) : 'N/A');
             $result['second'] = sprintf('%s (%s)', $data['amount'] / 100.0, $date);
         }
         $data = $info->getSfcoThirdPayment();
         if (!empty($data)) {
             $data = unserialize($data);
-            $date = preg_replace('/^([0-9]{2})([0-9]{2})([0-9]{4})$/', '$1/$2/$3', $data['date']);
+            $date = (isset($data['date']) ? preg_replace('/^([0-9]{2})([0-9]{2})([0-9]{4})$/', '$1/$2/$3', $data['date']) : 'N/A');
             $result['third'] = sprintf('%s (%s)', $data['amount'] / 100.0, $date);
         }
         return $result;